requirements.txt
Flask==1.1.1
google-cloud-datastore==1.10.0
setuptools==40.3.0
google-auth==1.11.2
requests==2.22.0